# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "University Admissions System API",
        "description": "RESTful API for managing university admissions, student enrollment, and academic records",
        "contact": {
            "email": "admin@university.edu"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API V1"
        }
    ],
    "paths": {
        "/api/v1/students/{student}/academic-records": {
            "get": {
                "tags": [
                    "Academic Records"
                ],
                "summary": "Get a student's academic records",
                "description": "Retrieve a list of academic records for a specific student.",
                "operationId": "1695a313a81a112508503850fe256d7d",
                "parameters": [
                    {
                        "name": "student",
                        "in": "path",
                        "description": "The ID of the student.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AcademicRecordResource"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Academic Records"
                ],
                "summary": "Create a new academic record",
                "description": "Create a new academic record for a student. Only administrators can create academic records.",
                "operationId": "9277acecd95bca24b717ed4c13c0d027",
                "parameters": [
                    {
                        "name": "student",
                        "in": "path",
                        "description": "The ID of the student.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreAcademicRecordRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Academic record created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AcademicRecordResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/academic-records/{academic_record}": {
            "get": {
                "tags": [
                    "Academic Records"
                ],
                "summary": "Get a specific academic record",
                "description": "Retrieve the details of a specific academic record.",
                "operationId": "bb908a1f0ec469e66d4380f7370966fe",
                "parameters": [
                    {
                        "name": "academic_record",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AcademicRecordResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Academic Records"
                ],
                "summary": "Update an academic record",
                "description": "Update an academic record. Only administrators can update academic records.",
                "operationId": "acbef71c9974f13cfa64cf79cbd33b42",
                "parameters": [
                    {
                        "name": "academic_record",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateAcademicRecordRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Academic record updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AcademicRecordResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Academic Records"
                ],
                "summary": "Delete an academic record",
                "description": "Delete an academic record. Only administrators can delete academic records.",
                "operationId": "28039e25049d3ab874642d02c00e8b24",
                "parameters": [
                    {
                        "name": "academic_record",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Academic record deleted successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "StoreAcademicRecordRequest": {
                "title": "Store Academic Record Request",
                "required": [
                    "institution_name",
                    "qualification_type",
                    "start_date",
                    "end_date",
                    "gpa"
                ],
                "properties": {
                    "institution_name": {
                        "description": "The name of the educational institution",
                        "type": "string",
                        "example": "University of California"
                    },
                    "qualification_type": {
                        "description": "The type of qualification or degree",
                        "type": "string",
                        "example": "Bachelor of Science"
                    },
                    "start_date": {
                        "description": "The start date of the academic program",
                        "type": "string",
                        "format": "date",
                        "example": "2020-09-01"
                    },
                    "end_date": {
                        "description": "The end date of the academic program",
                        "type": "string",
                        "format": "date",
                        "example": "2024-06-15"
                    },
                    "gpa": {
                        "description": "The Grade Point Average achieved (0.00-4.00)",
                        "type": "number",
                        "format": "float",
                        "example": 3.85
                    },
                    "transcript_url": {
                        "description": "URL to the transcript document",
                        "type": "string",
                        "example": "https://example.com/transcript.pdf",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "UpdateAcademicRecordRequest": {
                "title": "Update Academic Record Request",
                "properties": {
                    "institution_name": {
                        "description": "The name of the educational institution",
                        "type": "string",
                        "example": "University of California"
                    },
                    "qualification_type": {
                        "description": "The type of qualification or degree",
                        "type": "string",
                        "example": "Bachelor of Science"
                    },
                    "start_date": {
                        "description": "The start date of the academic program",
                        "type": "string",
                        "format": "date",
                        "example": "2020-09-01"
                    },
                    "end_date": {
                        "description": "The end date of the academic program",
                        "type": "string",
                        "format": "date",
                        "example": "2024-06-15"
                    },
                    "gpa": {
                        "description": "The Grade Point Average achieved (0.00-4.00)",
                        "type": "number",
                        "format": "float",
                        "example": 3.85
                    },
                    "transcript_url": {
                        "description": "URL to the transcript document",
                        "type": "string",
                        "example": "https://example.com/transcript.pdf",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "AcademicRecordResource": {
                "title": "Academic Record Resource",
                "description": "Represents an academic record for a student, such as their educational background and qualifications.",
                "properties": {
                    "id": {
                        "description": "The unique identifier for the academic record.",
                        "type": "integer",
                        "example": 1
                    },
                    "student_id": {
                        "description": "The ID of the student this record belongs to.",
                        "type": "integer",
                        "example": 1
                    },
                    "institution_name": {
                        "description": "The name of the educational institution.",
                        "type": "string",
                        "example": "University of California"
                    },
                    "qualification_type": {
                        "description": "The type of qualification or degree.",
                        "type": "string",
                        "example": "Bachelor of Science"
                    },
                    "start_date": {
                        "description": "The start date of the academic program.",
                        "type": "string",
                        "format": "date",
                        "example": "2020-09-01"
                    },
                    "end_date": {
                        "description": "The end date of the academic program.",
                        "type": "string",
                        "format": "date",
                        "example": "2024-06-15"
                    },
                    "gpa": {
                        "description": "The Grade Point Average achieved.",
                        "type": "number",
                        "format": "float",
                        "example": 3.85
                    },
                    "transcript_url": {
                        "description": "URL to the transcript document.",
                        "type": "string",
                        "example": "https://example.com/transcript.pdf",
                        "nullable": true
                    },
                    "verified": {
                        "description": "Whether the academic record has been verified.",
                        "type": "boolean",
                        "example": false
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}